Add paging and sorting support for custom queries.
In order to achieve paging and sorting support for custom queries,
this commit will refactor the query augmentation by introducing a query parser.
This feature needs to make assumptions on the query semantics.

If the query is intended for node- or relationship
entities, the first RETURN alias will be the anchor for sorting.

If the query asks for a path and has a with clause right before
the path, the with-clause will be used for sorting and pagination to
narrow down the path.

If no previous with-clause is present, one will be generated after the
path-match with node and relationship aliases. The first alias found
will be used for sorting.