2.0.1-SNAPSHOT: Extracted each driver into separate module. All driver jars now have test-scope
This has been done for licence compatibility reasons. It means that the drivers are no longer pulled
in transitively with the rest of the OGM modules. Consequently, if you are going to use 2.0.1-SNAPSHOT
you will need to explicitly declare which driver(s) you want to use in your project's pom.xml, e.g.

        <dependency>
            <groupId>org.neo4j</groupId>
            <artifactId>neo4j-ogm-http-driver</artifactId>
            <version>2.0.1-SNAPSHOT</version>
        </dependency>