Add trust strategy 'ACCEPT_UNSIGNED' for the HttpDriver
The Http driver now accepts the same configuration
settings as the Bolt driver, although not all
options are currently implemented.

This commit implements the following configuration settings:

- trust.strategy=ACCEPT_UNSIGNED
- connection_pool_size=xxx.

This allows users to configure the Http driver much more
easily than before.

Notable Changes:
1. The initialise() method on HttpDriver has been
removed because it did not work properly. Instead, a new
constructor has been implemented that accepts a user-defined
HttpClient instance.

2. The default HttpClient is now lazily-loaded. This is because
the configuration of the Driver occurs after it has been
instantiated. If a user has not explicitly provided an HttpClient
instance, one will be created the first time it is used from
the supplied configuration.