Neo4j-OGM is now (also) a set of OSGi bundles. The jar files still can be used normally. The only difference are some more META-INF/MANIFEST.MF entries that are relevant to OSGi containers.
neo4j-ogm-api is a real bundle
neo4j-ogm-compiler is a real bundle
neo4j-ogm-bolt-driver is a fragment to neo4j-ogm-api
neo4j-ogm-http-driver is a fragment to neo4j-ogm-api
neo4j-ogm-core is fragment to neo4j-ogm-api
neo4j-ogm-equinox-resolver is fragment to neo4j-ogm-api and introduces an additional
ResolverResolver for Equinox.

Some META-INF/services have been moved to the place of the actual provider.