Make rollback close bolt session when transaction is closed (#364)
Make rollback close bolt session when transaction is closed  `BoltTransaction` is the owner of neo4j driver's `Session` and is responsible for it's lifecycle. It's `#rollback()` method did not try to close session if native transaction was closed. However transaction and session are, to some extent, independent and both should always be closed to free up resources.  This commit makes `BoltTransaction#rollback()` always close session and adds tests for `#commit()` and `#rollback()`.  * Small cleanup   * removed usages of a deprecated method  * extracted conditional closing of session in a method  (cherry picked from commit 6751b2c)