Change graph id handling for new entities #381
Replace `System.identityHashCode` with decrementing counter.

This commit changes how org.neo4j.ogm.utils.EntityUtils.identity()
works.

When the id is present it is returned - same as before.

When the id is missing identityHashCode was used, which means multiple
calls returned same value. To keep this behaviour the id generated value
is written to the entity.

Code that depended on the null value of the id was changed to
< 0 and >= 0 comparisons.