Fix issue with empty (non null) collections
Happens only when session.query is used - it uses RestModelAdapter
under the hood. Convert arrays to iterables in the same way as
GraphModelAdapter does. Fixes #388.

Collections are stored as arrays. Empty collection is converted to
empty array, which does not have a type in cypher. When the collection
is retrieved the array is String[0].