Make possible to load a relationship referring to a non annotated abstract class. Fixes #404
Change request generation to ignore label of end node if the end node type
is a non annotated abstract class.

This also required to change the way how simple relationship values are written :
not according to target entity type but according to the collection type
which can be broader than specific target type.