GH-437 - Improve label determination and initialization of AutoIndexManger.
The AutoIndexManger now checks each registered entity if it needs to be indexed, either through declaration in itself or in a possible hierarchy.
Index fields are taken with priority from super classes to ensure indexes of the broadest range of nodes.

AutoIndexManger is no longer eagerly initiazized when AutoIndexMode is set to NONE to prevent opening a session on startup.
Tests around index creation have been revised in foresigt of removing some public methods of AutoIndexManger.

Queries are no longer executed if a label for a class or hierarchy cannot be computed (i.e. for abstract classes without @NodeEntity).

Co-authored-by: Jonathan D'Orleans <jonathan.dorleans@gmail.com>