Allow write protection of invidiual node properties
This change introduces #addWriteProtection on the Neo4jSession to register a predicate that determines whether certain aspects of Nodes should be read only or not. It also introduces the idea of a custom WriteProtectionStrategy, that allows for more fine grained control.

The feature is not on the Session itself to allow us some time to gather experience how it’s gonna be used and if its useful to a broader audience.

Co-authored-by: Jared Hancock <jarhanco@cisco.com>