Add infrastructure for CypherModification SPI
This introduces `org.neo4j.ogm.spi.CypherModificationProvider` that is used to bring in functions that consume the Cypher statements generated by OGM or passed through the Neo4j session in order to apply arbitrary modifications to them.
The mechanism uses Javas builtin SPI to load implementations of the above provider and it's ok to have more than one of them.

The changes also add `customProperties` to the OGM configuration to add arbitrary parameters (also dynamic suppliers of things) to the configuration. Those are passed on to the Cypher modification providers.