GH-518 - Track dynamic, composite properties.
This changes introduces the idea of a snapshot how a node was mapped to an entity in regard of dynamic properties.
So far dynamic properties of an entity are:

* All labels added through @Label (applies only to Nodes)
* All properties mapped with or through composite attribute converters (applies to Nodes and Relationships)

The new class EntitySnapshot replaces the previous LabelHistory and unifies these topics.

The new flow is clear on when labels are actually removed:
Before the change, the EntityGraphMapper had removePreviousLabelsIfRequired.
That method only stored the previous labels in the node builder.

Additional classes AbstractPropertyContainer and AbstractPropertyContainerBuilder have been introduced along with changes to PropertyContainerBuilder.
Those infrastructure allows a unified removal of labels and properties in ExistingNodeStatementBuilder and ExistingRelationshipStatementBuilder.

Tests have been added for both properties on Nodes and Relationships.

As a side effect, this change also allows "nulling" a location based on the old, non-native types.