GH-518 - Separate snapshots of node and relationship entities.
Relationships and entities retrieve their native ids from different buckets, so nodes and relationships may have the same native ids.
The original solution of „LabelHistory“ already created a situation where a relationship history would overwrite the one of a node with the same id.
However, that could not happened, as the previous solution was applied only to „things with a label field containing data“.
Now that we apply this to both nodes and relationships and all the time, the snapshots of entities and nodes must be separated.