Remove reference to transaction manager from driver.
The driver is the long living thing inside OGM. In any multithreaded
environment, it is dangerous to keep a transactionmanager like ours,
tied to a session and transaction and vice versa, here.

This commit replaces it with a supplier of factories for transactions
and decouples at least driver and transactionmanager.

This fixes especially the bookmark handling.

The bug fix cannot be done without interface changes.

This is essentially the same change as on 3.1, but with the added
improval that the transaction object is cleaned up after commit or
rollback as well.