GH-609 - Fix determination of visited nodes.
When a node is mapped with a depth of -1, all initial reachable nodes
will be stored in the visited objects with a horizon of -1 and therefore
will never appear as visited later on.

The fix is to treat each node visited in the first place as visited,
when asked for with a horizon of -1.

This fixes #609. Thanks @nithril and @Andy2003.