GH-627 - Turn off client side Cypher validation, request TX type as needed.
This turns off the client side Cypher parsing whether a query is read only or not, which fails in too many situations (i.e. properties containing „drop“ etc.)

It now requests a readOnly transaction when the flag is set to true and logs a warning, if the query contains keywords that might indicate a write query.

In addition to the corresponding change in 3.1 (7ed104211bb64b7de505370ac26ce4c0a9480858), this removes the deprecated query method, too.