GH-623 - Apply optimistic locking logic during merge, too.
Nodes with assigned ids (i.e. UUID based IDs or fully external IDs) are treated as new nodes, i.e. when loaded in one session and stored in another. OGM uses a MERGE statement for those. The optimistic locking clause is now added to that MERGE statement as well.

This fixes #623.