GH-653 - Improve support for Kotlin data classes.
This improves the support for Kotlin data classes. final fields are no longer excluded from the mapping. Instead only synthetic fields are excluded as that was the original intend of the exclusion.

A test is added to prove the working mapping, however the no-arg constructor kotlin-maven-plugin is needed due to the workings of OGM.

This closes #653.