Fix inconsistent usage of optimistic locking properties.
It was possible in some mapping scenarios that the saving of one start nodes having a rich relationship to several different endnodes would fail due to the deletion of relationships. The failure was caused by the deletion of one relationship removing the necessary information of the optimistic locking properties of the other relationships from the mapping context.