Add TestContainers and Neo4j 4.0 support.
Ensure that Bolt and HTTP will work with Neo4j 4.0.
Because Neo4j 4.0 will bring breaking API changes,
it is not easily possible to support embedded right now.

For the future it is planned to bring in 4.0 support for embedded
but this is not in the scope of this commit.