GH-684 - Fix possible NPE in IdentityMap.
The identity map used the wrong map to look up the expected hash value when both a node and relationship with the same id have been loaded and a NPE happened when casting literal Null to long. This has been fixed by first selecting the correct hash.

The error first appeared in the performance optimizations in #576, #577 and #579.