GH-345 - Decouple filter and filter functions.
This removes the tight coupling of filters and filter functions. The property name that is filtered will now be injected into the filter function and so will be an operator to generate a unique parameter name. This will break API for users running their own filter functions, but make things a lot simpelr in the future.

Also, the API of the filter has been reduced, so that is now clearer which constructor to use.