GH-696 - Fix resolving of type descriptor in Kotlin collections.
The Kotlin compiler uses an upper bounded wildcard type for Kotlin based node entities having Kotlin collections of Java based node entities. The previous resolving algorithm resolves this as an object, because it doesn’t do a check for wildcard types.

We only add support for wildcards (for now) for usage in Kotlin, as there are explicit tests in place checking for `java.lang.Object` as type descriptor and check wether the containing type is a Kotlin class.

This fixes #696.