GH-657 - Fix deletion of bidirectional, undirected relationships.
This change introduces
* immutable MappedRelationship
* A tracker root level objects

The tracker for root level objects is used to decide, wether a pending, undirected relationship needs to be mapped at the end of a transaction or not. The immutable mapped relationships are somewhat cosmetic, but makes it easier to understand the mapping process.

This is a backport of the corresponding commit in 3.2.x.