GH-613 - Fix incomplete removal of stale relationships and nodes from context.
The deleteObsoleteRelationships() method now marks all nodes reachable from stale relationship stale as well. If a node was only reachable via one of the original relationships, than it can actually be stale and will be removed.

If it get’s removed, it will in turn mark more relationships as stale, which also will be removed.

This fixes #613.