GH-704 - Improve building of class hierarchies.
This is a backport of 4986061351db5a9f855aaacec62dabc6223a0b46, which
need a bit more work, as the class scanning structure was changed
substantially in 3.2. The main idea however is still the same: By
loading a class, all of its superclasses will be loaded anyway and we
don't need to do this recursively and can omit the error prone hydration
of things.