GH-551 - Make SingleUseEntityMapper aware of nested objects.
This change allows for returning nested objects, that are actually part of the scanned classes, to be returned from the `SingleUseEntityMapper` when the target type of a collection fits. Otherwise the `SingleUseEntityMapper` would silently put `Map<String, Object>` instances in collections, which would fail with a late `ClassCastException` when items of the collection property are eventually accessed.

While implementing the functionality, it became obvious that `GraphEntityMapper` could be optimized by not looking up the field info several times.

The commit fixes #551.