GH-651 - Use all mapped labels when querying domain objects.
During loadOneById, loadAllByIds and loadAll all labels shall be added to a query, thus making indexes defined at parent classes actually usable.

This commit adds a dedicated method to Neo4jSession to determine a label, multiple labels or a type. The labels need to be pre joined with a colon and half the separator. See the corresponding comment. The `SchemaNodeLoadClauseBuilder` had to be adapted for that.

This is an opt-out feature for 4.0.

The commit also makes the computation of static labels a bit less repetitive.

Co-Authored-By: Gerrit Meier <meistermeier@gmail.com>