GH-651 - Use all mapped labels when querying domain objects.
During loadOneById, loadAllByIds and loadAll all labels shall be added to a query, thus making indexes defined at parent classes actually usable.

This commit adds a dedicated method to Neo4jSession to determine a label, multiple labels or a type. The labels need to be pre joined with a colon and half the separator. See the corresponding comment. The `SchemaNodeLoadClauseBuilder` had to be adapted for that.

This commit is a backport of b90d4136c178015cb3adf4401ca7fe9376ea531a  and makes it an opt-in feature for the 3.x line. The feature can be enabled via a the configuration property `use-strict-querying` or by using a dedicated `SessionFactory` constructor.