GH-670 - Fix inconsistend behaviour when querying parent classes.
OGM allows to query for broader types than actually be returned. For example, one can query a Person and get back concrete Pupil and Teacher nodes.

When a new node with relationships defined at the parent class is stored and queried immediatly, the relationships are filled.
This works due to our caches.
If the session is cleared or the object is queried with the correct fetch depth in another session, the relationships are not filled.
This is because the schema builder only generates pattern comprehensions for a concrete given class.

While this is technically not wrong, it should definitiv be consistent with the ways we allow querying: So either just returning the broadest concrete type possible, or the correct type with the relationship filled.

We settled for the later one and generate now more pattern comprehensions in scenarios where a parent class is queried but child classes defines additional relationships.
We do this recursivly for all child classes.

This can lead to slower queries for broad types, but than this must be fixed at the querying side.

This commit moves all query generation tests to the test module, so that the test domain classes doesn’t need to be repeated and therefore is rather large.

This fixes #670.