GH-666 - Check for contradicting annotations before mapping possible related nodes.
The entity access manager tries several ways to find a field for mapping relationships. Usually, more than 3 tries are necessary for custom queries. The last approach just checks if there is any field of the same type as a currently mapped relationship and just dumps the relationships content into it.

This is bad in case when the field type matches but there’s actually an annotation contradicting it.

This commit makes the entity access manager check for contradicting annotations and if there is one, doesn’t fill in the object.

This is a backport of 0a9a921ac4d562df2e99a855f662ba2d41c031de.