GH-718 - Recursively traverse variable length relationship patterns.
This change fixes the inconsistent behavior of the `org.neo4j.ogm.result.adapter.RestModelAdapter` that didn’t traverse nested things as the `org.neo4j.ogm.result.adapter.GraphModelAdapter` does. Would a fixed length pattern had been used in #718 the mapping would have worked as expected.

However, it is actual the point to collect all of the relationship entities of the given point, so we have to traverse the collection of relationships we get back.

This fixes #718.