GH-727 - Return unmapped relationship models.
In case a relationship model can’t be mapped (either because the result set doesn’t contain the start or end node or the relationship is not a relationship entity), don’t drop it from the result set but return the unmapped model, as it was done in 3.1.x.

This fixes #727.