GH-728 - Improve class loading mechanism.
Neo4j-OGM used the current Threads class loader for loading domain classes and a slightly different approach for the native types since 3.2.

While that works in most scenarios, there are some systems supporting dynamic plugins etc, that initialize the plugin with a different classloader but not in a different thread with a dedicated thread context class loader. In those scenarios we must fall back to use the class loader used for OGM itself.

This commit introduces a delegating classloader that does exactly that. To keep Neo4j-OGMs behaviour consistent with 3.1 and 3.2, it consulates the thread class loader first, than any other delegate.

The delegating class loader provides the opportunity to make this configurable if necessary.

Furthermore, loading of resources during configuration phase of Neo4j-OGM has been consolidated to use the same mechanism.

Unrelated polishing: Update Classgraph to 4.8.60 and reopen a session during test.

Co-authored-by: Gerrit Meier <meistermeier@gmail.com>