GH-743 - Fix IllegalArgumentException on first reload with DevTools.
This rolls back the solution from #728, so that the first restart of
Spring Boots restart class loader doesn't fail anymore.

The problem with the delegating class loader is that the VM knows it has
loaded the classes mapped once, regardless the fact that it's underlying
Thread context class loader changed. As there is no way to overwrite
that behavior, we cannot keep the delegating class loader solution.

For #728, we will provide a configurable solution.

This closes #743.