GH-748 - Fix return type for known entity classes.
Use null instead of an empty list when the target isn’t a collection as well. This verifies the indirect illegal argument exception in case target property and query won’t fit but keeps the lenient behaviour in place when a query returning collections actually returns single valued collections.

The scopes of the logback dependencies have been clarified.

Thanks to @nioertel for reporting the issue and helping to solve it.

This fixes #748.