GH-782 - Prevent NPE when 1:n relationships contains null elements.
This introduces an `InvalidRelationshipTargetException` which is thrown when a user tries to persist a domain object holding a 1:n reference in a collection containing one or more `null` elements. This prevents a `NullpointerException` and fixes #781. 

The change also reduces the number of branches in `EntityGraphMapper#mapEntityReferences` as we can reuse an `iterableOf` helper introduced for the `SingleUseEntityMapper`. Making this helper available as a utils method, we could also remove similar usage in other areas as a side effect.