Remove getEnum-API.
This removes the new API on the AnnotationInfo. This API was very well done and tested, but it increases our surface again and we don’t want that. It’s needed in two places and it’s ok to duplicate it the instantiation there.

There’s one signifacant change we’re we check whether the attribute is an enum or not.  For comparision: https://github.com/neo4j/neo4j-ogm/pull/644