Further polishing.
This introduces some aliases on the `@Relationship` for the `Direction`
constants. These provide some compatibility for users relying on the
previous string constants in their declarations.

Furthermore, the availibility of the enum shows some points that should
be refactored all too painfull, especially in `ClassInfo` which has been
done.

The rest is minor polishing.