Unwrap composite keys during load and save.
Primary keys having a `CompositeAttributeConverter` have not been taking into consideration when saving and loading: The map would be used as "the" value for the id. That would might work when the map would be converted by the drivers, but the queries would not match anything due to the fact that no keys in the map are compared.

This commit changes this by creating an inline cypher map from the properties that are matched on during load and written during save as well.

This fixes #789.