Make the AutoIndexManager aware of decomposed fields.
This allows properties in @CompositeIndex to be prefixed with `fieldName.` so that the auto index manager can recognize indexes belonging to fields that are decomposed by a map composite converter.

We also log a big warning in case a field is to be decomposed but doesn’t carry a composite index but a normal one.