GH-790 - Treat composite keys correct during load and save operations.
This change has a couple of aspects:

= Unwrap composite keys during load and save.

Primary keys having a `CompositeAttributeConverter` have not been taking into consideration when saving and loading: The map would be used as "the" value for the id. That would might work when the map would be converted by the drivers, but the queries would not match anything due to the fact that no keys in the map are compared.

This commit changes this by creating an inline cypher map from the properties that are matched on during load and written during save as well.

This fixes #789.

= Make the AutoIndexManager aware of decomposed fields.

This allows properties in @CompositeIndex to be prefixed with `fieldName.` so that the auto index manager can recognize indexes belonging to fields that are decomposed by a map composite converter.

We also log a big warning in case a field is to be decomposed but doesn’t carry a composite index but a normal one.