GH-793 - Fix reading of Neo4j literal byte[] arrays.
We optimized the previous `convertToIterable` (now `iterableOf` in `CollectionUtils` to actually return an iterable. In the case of byte arrays, the `EntityAccessManager` casts that `Iterable` into a collection to coerce it with existing values. That lead to the `ClassCastException` mentioned in #791.

The issue needs several fixes:
- `EntityAccessManager` must only work on `Iterables`
- The `CollectionUtils` must treat an array of primitive types separately. It cannot be just casted into an `Object[]` array.
- The `BaseAdapter` class is now clear about the fact that it turns arrays into collections and not only iterables. If it doesn't do this, several things breaks downstream.

This change also introduces two NoOp-Converter (`NoOpByteArrayConverter` and `NoOpWrappedByteArrayConverter`) that allow users to circumvent the conversion to Base64 arrays.

We can read Base64 encoded into byte[] arrays from HTTP, but not write byte[] arrays over HTTP as the HTTP endpoint doesn’t support byte[] as parameter.

This closes #791.