Prevent OR’ing of filters containing nested paths.
A filter with nested paths will result in several matches chained together. Those matches are nessary to have the nodes holding the properties we want to filter on.

While at first it may seem trivial to think that adding a nested filter with its own match by an OR just doing an OPTIONAL MATCH, this is wrong, as the first part of the OR would need to be marked as OPTIONAL as well.

In the end trying to do so, would create very brittle Cypher statements.

Instead of, this commit improves the current checks into a single place and adds a couple of tests for it.

A query is not created when any of the filters in the chain is nested AND any other is to be chained with OR (when there is actually more than 1 filter).