GH-800 - Improve nested filtering.
This change allows having OR’s in nested filters targeting the same entity. Thus, the entity itself works as a logical group and can itself by combined with filters on other properties with AND.

A combination of filters with OR containing one nested filter is still not allowed.

This commit makes the `NodeQueryBuilder` collapse deeply nested filters the same way as it does with normal ones.