Improve handling of unknown generic types.
This improves the `SingleUseEntityMapper` to check the returned value for being something like a collection in case we couldn’t determine the generics of a wrapper class. In case the of Collections, this allows for a correct assigment.

Generic arrays are not detectable with that approach.

So something like

```
public class GenericQueryResultWrapper<T> {

    T result;
}
```

will work for `T` being a known entity type or a collection thereof, but not for an array of known entities.