GH-831 - Avoid early population of externally generated ID fields.
This introduces `optionalNativeId` on the mapping context that does not trigger any id generating strategies in contrast tp `nativeId`. This method is than used especially in the eventing system so that firing a pre-save event contains the correct `isNew`state.

Furthermore it is now used in checking dirtiness and in the delete events, so that those don’t modify the object states without any added benefit.