GH-846 - Fix determination of target graph property type.
* Fix determination of target graph property type.

The algorithm that determines the graph property type is faulty: It is not enough to retrieve the declared methods of a class and filter out the synthetic ones to get the most specific method implementation or overwrite.

The main goal of that code in FieldInfo is about getting the actual return type without trying to figure this out from generic parameters. This is done now in multiple steps:

If we find more than one but one none synthetic, we use this (same as before).
If all we finde are synthetic, we keep those declared on the actual converter class.
If those are more than one, we take the one not returning generic Object.

This fixes #845.