GH-854 - Avoid recursive traversal for computing events.
This change avoids the recursive depth first iteration of nodes when visiting and testing them if we should fire a pre-save event.

To make this work, dirty must be fixed in two places:
- Checking for relationship entities trigges an ID creation to early (compare GH-831, 18059993d0ce9450cc62994c2e44d68d07466236)
- we must ensure that we check only for existing previous relationships that are actually touched by the object whos dirtyness is checked

Furthermore, the change avoids a lot of checks for the native id of the changed object and treats collections now correctly.

This closes #854.