GH-851 - Check if properties belong to the same entity but different relationships.
When filters (nested or „deep“ nested) are used on properties of the same entity but with a different relationship, they are folded together. While an OR query would retrieve some results (often times to many), an AND query would retrieve no results. This changes checks the property if it points to the same relationship and if so, generates multiple match clauses.

This fixes #851.