GH-875 - Use all available types when retrieving polymorphic mapped relationships.
This adds an additional collection to the `Node` and its implementation in the schema, containing the collection of types which a specific relationship is using. The list will contain multiple types only when one relationship is equal in the field (attribute) name and both the type of start and other node.

Multiple types will be queried using the Cypher `|` operator.

This fixes #875.