Adapt auto-index manager to changes in Neo4j 4.3.
This change makes the Auto-Index manager compatible with Neo4j 4.3:

- It ignores the index type `LOOKUP` which defines the label / relationship type lookup
- It adapts to the new assertion of existing properties during read (`IS NOT NULL` instead of `exists`)

For a future version the auto index manager must be adapted to use the new syntax, but that would require such deep changes, that a rewrite into an interface plus a 3.x and 4.x implementation is necessary. Adding more condition to the current version is pointless.