Migrate to Neo4j 4.x.
This brings the main branch of this repository to Neo4j 4.x.
Neo4j 4.x has been supported via a 4.x driver for quite a while when using the `Bolt` transport.
However, bringing up a Neo4j 4.x embedded was not possible. 
This commit changes it but also drops support for all 3.x versions for this branch.
With this, the additional native types projects can go away as well, as they are not needed as separate modules anymore.
All supported Neo4j versions support.

A release of this branch is not planned yet.