fix: Don't flatten collections of known entities.
This fixes #902 in a similar fashion to what
0ba80b1961a06781707b03e19bed4a41b8284998 did for #821.

While #821 was fixed for non-domain objects as the `handle` method of
the `RestModelMapper.ResultRowBuilder` calls `convertToTargetContainer
which itself works recursively on non-domain objects, such an implicit
fix wasn't applied to domain objects.

As the domain objects are all flattened into the in-memory graph model
the hierarchy (parents and siblings) of nested lists must be stored
inside the `RestModelMapper`. This is achivied by a `Map<String,
List<String>>`, basically a nested list of pointers.

This map is used to recreate the shape of the original query.

This had some affects on the `SingleUseEntityMapper`: It may run into a
list of already mapped objects now, so those need some special
treatment, similar to queries such as mentioned in #737 (especially
queries containing and returning pattern comprehension without
extracting nodes and relationsships): They will work as before and
contain the mapped objects, but in the same fashion the pattern
comprehension produced them: In a list.