GH-903 - Increment only version properties of changed relationship entities.
The fix is to use an increment of 0 in the statements handling existing relationship entities. While I tried at first to exclude to corresponding builders in `org.neo4j.ogm.context.EntityGraphMapper#getRelationshipBuilder` (and with them the edges and eventually the statements) from the list of statements to execute, we didn’t succeed with that. The `PizzaIntegrationTest`, more specifically, the `shouldBeAbleToSaveAndRetrieveFullyLoadedPizza` became flaky with that.

I have no clue what the actual tests meaning is, but I assume the underlying cause is the fact the the context won’t get updated with all the affected entities (the test uses a couple of different sessions). That means: A fully hydrated and changed object is updated in one session, which may or may not have the information about the content of the local graph. When the statements of unchanged entities are excluded, they won’t never be returned from the datatbase, hence not make their way into the session cache and therefore will never update the update in question.

This fixes #903.