refactor: Add an extension API for custom transaction managers. (#935)
This change does the following

1. Clean up the Transaction manager api
The previous state mixed responsibilities: whether a transaction could be committed or not, rolled back or not etc. was scattered across the transaction manager and the transaction itself. The same thinking was visible for the actual work of committing or rolling back things. This has been separated. Until this change, the API was not declared in any shape or form a public API. So the change is reasonable safe todo. The transaction interface itself stays stable from a callers perspective (not from an implementors though, but it is only intended to be implemented by us anyhow).

2. Add a base implementation
The transaction managed follows a certain OGM logic in context with the session as well as possible extendibility of sessions (kind of a poor mans nested transaction). While the latter is seldom used, the former is important. Especially clearing of newly registered objects in a session is paramount. For that purpose an abstract base implementation of the transaction manager has been added. The sensible methods are all final. It exposes however `openOrExtend` and `removeIfCurrent` as abstract methods, both parameterised with actions that we provide internally. So all that's left to do for issues like mentioned in #934 that needs different ways than thread locals to attach sessions to threads, those are the entry point

4. Provide extension points on the concrete classes
`SessionFactory` allows now configuration of transaction manager factories and `Neo4jSession` allows access to any transaction manager configured.

This closes #934.