GH-951 - Direct relationships can overrule relationship entities.
If the source model has a relationship of a specific type defined that
is also covered by a relationship entity, Neo4j-OGM would eagerly
try to map everything to this entity before it notices that the source
is not using the entity and does not put in the related node because
the types are mismatching.
This commit will fix this behaviour while keeping already existing functionality in place as it is.