Bump default driver version to latest 4.4.x series.
While we communicated clearly that bumping from the 4.0.x driver to any higher 4.m.x driver was never a problem, we decided it’s time to do the bump ourselves.

This has a couple of consequences

- It’s a minor bump, not a patch due to semver requirements. There are breaking API changes on our side though
- The Neo4j 4.4.x driver dropped support for older Neo4j versions, especially 3.2 and 3.3. While it may support 3.4, we decided to go with it’s official line and support only 3.5. In addition, as Neo4j 4.4.x is the LTS 4 release of the database, we will test only against 3.5 and 4.4
- That means that we will only fix things that happen against older databases on a case by case basis

In general, bumping to any Neo4j-OGM 3.3 release from 3.2.42 is just a drop-in replacement on your side if you already on Neo4j 4.4 release.

This change does not bump the minimum required Java version and the Neo4j-OGM 3.m series stays compatible with Java 8.