feat: Make sure that the usage of Java records does not end in tears.
Be warned that this is not something we recommend doing, as records are very much immutable.
So it might give you a good results for read-only entities, but mix experience when updating.
The only scenario that will work here are entities with assigned ids and really careful session handling.
You have been warned.